+++ packages/design-system-react/src/components/Tag/Tag.test.tsx
@@ -1,8 +1,7 @@
-import { TagSeverity } from '@metamask/design-system-shared';
+import { IconName, TagSeverity } from '@metamask/design-system-shared';
 import { render, screen } from '@testing-library/react';
 import React from 'react';
 
-import { IconName } from '../../types';
 import { TextColor } from '../Text';
 
 import { Tag } from './Tag';