+++ packages/design-system-shared/src/types/Icon/Icon.types.ts
@@ -313,6 +313,7 @@ export const IconName = {
   Trash: 'Trash',
   TrendDown: 'TrendDown',
   TrendUp: 'TrendUp',
+  Trophy: 'Trophy',
   Undo: 'Undo',
   Unfold: 'Unfold',
   UnlockedFilled: 'UnlockedFilled',